Refactoring utilities file into component/x + create get-template file for code share
