+++ components/pure-component.js
@@ -1,7 +1,7 @@
-const Utilities = require("./utilities");
+const getTemplate = require("./get-template");
 
 module.exports = PureComponent;
 
 function PureComponent(rootDirectory, name, hasCSS) {
-  return Utilities.getTemplate(rootDirectory, name, hasCSS, "pure-component");
+  return getTemplate(rootDirectory, name, hasCSS, "pure-component");
 }
